+++ aiohttp/web_reqrep.py
@@ -17,10 +17,9 @@ from urllib.parse import parse_qsl, unquote, urlsplit
 from multidict import CIMultiDict, CIMultiDictProxy, MultiDict, MultiDictProxy
 
 from . import hdrs
-from .helpers import _sentinel, reify
-from .protocol import RawRequestMessage
+from .helpers import reify, sentinel
 from .protocol import Response as ResponseImpl
-from .protocol import HttpVersion10, HttpVersion11
+from .protocol import HttpVersion10, HttpVersion11, RawRequestMessage
 from .streams import EOF_MARKER
 
 __all__ = (
@@ -33,7 +32,7 @@ class HeadersMixin:
 
     _content_type = None
     _content_dict = None
-    _stored_content_type = _sentinel
+    _stored_content_type = sentinel
 
     def _parse_content_type(self, raw):
         self._stored_content_type = raw
@@ -445,6 +444,8 @@ class StreamResponse(HeadersMixin):
 
         if headers is not None:
             self._headers.extend(headers)
+        self._parse_content_type(self._headers.get(hdrs.CONTENT_TYPE))
+        self._generate_content_type_header()
 
     def _copy_cookies(self):
         for cookie in self._cookies.values():
@@ -624,8 +625,7 @@ class StreamResponse(HeadersMixin):
     @last_modified.setter
     def last_modified(self, value):
         if value is None:
-            if hdrs.LAST_MODIFIED in self.headers:
-                del self.headers[hdrs.LAST_MODIFIED]
+            self.headers.pop(hdrs.LAST_MODIFIED, None)
         elif isinstance(value, (int, float)):
             self.headers[hdrs.LAST_MODIFIED] = time.strftime(
                 "%a, %d %b %Y %H:%M:%S GMT", time.gmtime(math.ceil(value)))
@@ -677,7 +677,7 @@ class StreamResponse(HeadersMixin):
         if self._resp_impl is not None:
             if self._req is not request:
                 raise RuntimeError(
-                    'Response has been started with different request.')
+                    "Response has been started with different request.")
             else:
                 return self._resp_impl
         else:
@@ -756,7 +756,7 @@ class StreamResponse(HeadersMixin):
 
     def write(self, data):
         assert isinstance(data, (bytes, bytearray, memoryview)), \
-            'data argument must be byte-ish (%r)' % type(data)
+            "data argument must be byte-ish (%r)" % type(data)
 
         if self._eof_sent:
             raise RuntimeError("Cannot call write() after write_eof()")
@@ -798,49 +798,55 @@ class Response(StreamResponse):
     def __init__(self, *, body=None, status=200,
                  reason=None, text=None, headers=None, content_type=None,
                  charset=None):
-        super().__init__(status=status, reason=reason, headers=headers)
-        self.set_tcp_cork(True)
-
         if body is not None and text is not None:
-            raise ValueError("body and text are not allowed together.")
+            raise ValueError("body and text are not allowed together")
+
+        if headers is None:
+            headers = CIMultiDict()
+        elif not isinstance(headers, (CIMultiDict, CIMultiDictProxy)):
+            headers = CIMultiDict(headers)
+
+        if content_type is not None and ";" in content_type:
+            raise ValueError("charset must not be in content_type "
+                             "argument")
 
         if text is not None:
-            if hdrs.CONTENT_TYPE not in self.headers:
+            if hdrs.CONTENT_TYPE in headers:
+                if content_type or charset:
+                    raise ValueError("passing both Content-Type header and "
+                                     "content_type or charset params "
+                                     "is forbidden")
+            else:
                 # fast path for filling headers
                 if not isinstance(text, str):
-                    raise TypeError('text argument must be str (%r)' %
+                    raise TypeError("text argument must be str (%r)" %
                                     type(text))
                 if content_type is None:
                     content_type = 'text/plain'
-                elif ";" in content_type:
-                    raise ValueError('charset must not be in content_type '
-                                     'argument')
-                charset = charset or 'utf-8'
-                self.headers[hdrs.CONTENT_TYPE] = (
-                    content_type + '; charset=%s' % charset)
-                self._content_type = content_type
-                self._content_dict = {'charset': charset}
-                self.body = text.encode(charset)
-            else:
-                self.text = text
-                if content_type or charset:
-                    raise ValueError("Passing both Content-Type header and "
-                                     "content_type or charset params "
-                                     "is forbidden")
+                if charset is None:
+                    charset = 'utf-8'
+                headers[hdrs.CONTENT_TYPE] = (
+                    content_type + '; charset=' + charset)
+                body = text.encode(charset)
+                text = None
         else:
-            if hdrs.CONTENT_TYPE in self.headers:
-                if content_type or charset:
-                    raise ValueError("Passing both Content-Type header and "
+            if hdrs.CONTENT_TYPE in headers:
+                if content_type is not None or charset is not None:
+                    raise ValueError("passing both Content-Type header and "
                                      "content_type or charset params "
                                      "is forbidden")
-            if content_type:
-                self.content_type = content_type
-            if charset:
-                self.charset = charset
-            if body is not None:
-                self.body = body
             else:
-                self.body = None
+                if content_type is not None:
+                    if charset is not None:
+                        content_type += '; charset=' + charset
+                    headers[hdrs.CONTENT_TYPE] = content_type
+
+        super().__init__(status=status, reason=reason, headers=headers)
+        self.set_tcp_cork(True)
+        if text is not None:
+            self.text = text
+        else:
+            self.body = body
 
     @property
     def body(self):
@@ -849,7 +855,7 @@ class Response(StreamResponse):
     @body.setter
     def body(self, body):
         if body is not None and not isinstance(body, bytes):
-            raise TypeError('body argument must be bytes (%r)' % type(body))
+            raise TypeError("body argument must be bytes (%r)" % type(body))
         self._body = body
         if body is not None:
             self.content_length = len(body)
@@ -865,7 +871,7 @@ class Response(StreamResponse):
     @text.setter
     def text(self, text):
         if text is not None and not isinstance(text, str):
-            raise TypeError('text argument must be str (%r)' % type(text))
+            raise TypeError("text argument must be str (%r)" % type(text))
 
         if self.content_type == 'application/octet-stream':
             self.content_type = 'text/plain'
@@ -874,28 +880,26 @@ class Response(StreamResponse):
 
         self.body = text.encode(self.charset)
 
-    def should_send_body(self):
-        return (self._req.method != hdrs.METH_HEAD and
-                self._status not in [204, 304])
-
     @asyncio.coroutine
     def write_eof(self):
         try:
             body = self._body
-            if body is not None and self.should_send_body():
+            if (body is not None and
+                    self._req.method != hdrs.METH_HEAD and
+                    self._status not in [204, 304]):
                 self.write(body)
         finally:
             self.set_tcp_nodelay(True)
         yield from super().write_eof()
 
 
-def json_response(data=_sentinel, *, text=None, body=None, status=200,
+def json_response(data=sentinel, *, text=None, body=None, status=200,
                   reason=None, headers=None, content_type='application/json',
                   dumps=json.dumps):
-    if data is not _sentinel:
+    if data is not sentinel:
         if text or body:
             raise ValueError(
-                'only one of data, text, or body should be specified'
+                "only one of data, text, or body should be specified"
             )
         else:
             text = dumps(data)