+++ aiohttp/web_ws.py
@@ -2,11 +2,13 @@ import asyncio
 import json
 import sys
 import warnings
+from collections import namedtuple
 
 from . import Timeout, hdrs
 from ._ws_impl import (CLOSED_MESSAGE, WebSocketError, WSMessage, WSMsgType,
                        do_handshake)
 from .errors import ClientDisconnectedError, HttpProcessingError
+from .helpers import _decorate_aiter
 from .web_exceptions import (HTTPBadRequest, HTTPInternalServerError,
                              HTTPMethodNotAllowed)
 from .web_reqrep import StreamResponse
@@ -14,9 +16,13 @@ from .web_reqrep import StreamResponse
 __all__ = ('WebSocketResponse',)
 
 PY_35 = sys.version_info >= (3, 5)
-PY_352 = sys.version_info >= (3, 5, 2)
 
 THRESHOLD_CONNLOST_ACCESS = 5
+
+
+class WebSocketReady(namedtuple('WebSocketReady', 'ok protocol')):
+    def __bool__(self):
+        return self.ok
 
 
 class WebSocketResponse(StreamResponse):
@@ -98,9 +104,9 @@ class WebSocketResponse(StreamResponse):
                 request.method, request.headers, request.transport,
                 self._protocols)
         except HttpProcessingError:
-            return False, None
+            return WebSocketReady(False, None)
         else:
-            return True, protocol
+            return WebSocketReady(True, protocol)
 
     def can_start(self, request):
         warnings.warn('use .can_prepare(request) instead', DeprecationWarning)
@@ -205,6 +211,10 @@ class WebSocketResponse(StreamResponse):
                 if msg.type == WSMsgType.CLOSE:
                     self._close_code = msg.data
                     return True
+
+            self._close_code = 1006
+            self._exception = asyncio.TimeoutError()
+            return True
         else:
             return False
 
@@ -249,13 +259,12 @@ class WebSocketResponse(StreamResponse):
                     if not self._closed and self._autoclose:
                         yield from self.close()
                     return msg
-                elif not self._closed:
-                    if msg.type == WSMsgType.PING and self._autoping:
-                        self.pong(msg.data)
-                    elif msg.type == WSMsgType.PONG and self._autoping:
-                        continue
-                    else:
-                        return msg
+                if msg.type == WSMsgType.PING and self._autoping:
+                    self.pong(msg.data)
+                elif msg.type == WSMsgType.PONG and self._autoping:
+                    continue
+                else:
+                    return msg
         finally:
             self._waiting = False
 
@@ -293,13 +302,9 @@ class WebSocketResponse(StreamResponse):
         raise RuntimeError("Cannot call .write() for websocket")
 
     if PY_35:
-        if PY_352:
-            def __aiter__(self):
-                return self
-        else:
-            @asyncio.coroutine
-            def __aiter__(self):
-                return self
+        @_decorate_aiter
+        def __aiter__(self):
+            return self
 
         @asyncio.coroutine
         def __anext__(self):