+++ tests/test_py35/test_resp.py
@@ -7,7 +7,6 @@ import aiohttp
 from aiohttp import web
 from aiohttp.client import _RequestContextManager
 
-
 async def test_await(create_server, loop):
 
     async def handler(request):